add js sitemap
